Fix: api cart, wishlist and order

+++ ZumZumFood.Application/Services/CartService.cs
@@ -1,4 +1,7 @@
-﻿namespace ZumZumFood.Application.Services
+﻿using StackExchange.Redis;
+using ZumZumFood.Domain.Entities;
+
+namespace ZumZumFood.Application.Services
 {
     public class CartService : ICartService
     {
@@ -28,6 +31,14 @@
                                            .ThenInclude(p => p.Category)
                                            .Include(x => x.Product)
                                            .ThenInclude(p => p.Brand)
+                                           .Include(d => d.ComboProduct)
+                                           .ThenInclude(co => co.Combo)
+                                           .Include(d => d.ComboProduct)
+                                           .ThenInclude(co => co.Product)
+                                           .ThenInclude(p => p.Category)
+                                           .Include(d => d.ComboProduct)
+                                           .ThenInclude(co => co.Product)
+                                           .ThenInclude(p => p.Brand)
                 );
                 var data = dataQuery;
                 var cart = dataQuery.FirstOrDefault();
@@ -53,9 +64,10 @@
                         DateOfBirth = cart.User.DateOfBirth.HasValue ? cart.User.DateOfBirth.Value.ToString("dd-MM-yyyy HH:mm:ss") : null,
                         Nationality = cart.User.Nationality,
                     },
-                    Products = data.Select(p => new ProductDTO
+                    Products = data.Where(x => x.ProductId != null) // Only filter for items with a non-null ProductId
+                    .Select(p => new ProductDTO
                     {
-                        ProductId = p.ProductId,
+                        ProductId = (int)p.ProductId,
                         Name = p.Product.Name,
                         Slug = p.Product.Slug,
                         Image = p.Product.Image,
@@ -71,6 +83,48 @@
                         DeleteBy = p.Product.DeleteBy,
                         DeleteDate = p.Product.DeleteDate.HasValue ? p.Product.DeleteDate.Value.ToString("dd-MM-yyyy HH:mm:ss") : null,
                         DeleteFlag = p.Product.DeleteFlag,
+                    }).ToList(),
+                    Combos = data.Where(x => x.ComboProductId != null) // Only filter for items with a non-null ProductId
+                    .Select(c => new ComboDTO
+                    {
+                        ComboId = c.ComboProduct.Combo.ComboId,
+                        Name = c.ComboProduct.Combo.Name,
+                        Image = c.ComboProduct.Combo.Image,
+                        Price = c.ComboProduct.Combo.Price,
+                        IsActive = c.ComboProduct.Combo.IsActive,
+                        Description = c.ComboProduct.Combo.Description,
+                        CreateBy = c.ComboProduct.Combo.CreateBy,
+                        CreateDate = c.ComboProduct.Combo.CreateDate.HasValue ? c.ComboProduct.Combo.CreateDate.Value.ToString("dd-MM-yyyy HH:mm:ss") : null,
+                        UpdateBy = c.ComboProduct.Combo.UpdateBy,
+                        UpdateDate = c.ComboProduct.Combo.UpdateDate.HasValue ? c.ComboProduct.Combo.UpdateDate.Value.ToString("dd-MM-yyyy HH:mm:ss") : null,
+                        DeleteBy = c.ComboProduct.Combo.DeleteBy,
+                        DeleteDate = c.ComboProduct.Combo.DeleteDate.HasValue ? c.ComboProduct.Combo.DeleteDate.Value.ToString("dd-MM-yyyy HH:mm:ss") : null,
+                        DeleteFlag = c.ComboProduct.Combo.DeleteFlag,
+                        Products = data
+                        .Where(d => d.ComboProduct?.Product != null) 
+                        .Where(d => d.ComboProduct.ProductId == c.ComboProduct.ProductId) 
+                        .Select(d => new ProductDTO
+                        {
+                            ProductId = d.ComboProduct.Product.ProductId,
+                            Name = d.ComboProduct.Product.Name,
+                            Slug = d.ComboProduct.Product.Slug,
+                            Image = d.ComboProduct.Product.Image,
+                            Price = d.ComboProduct.Product.Price,
+                            Discount = d.ComboProduct.Product.Discount,
+                            IsActive = d.ComboProduct.Product.IsActive,
+                            Description = d.ComboProduct.Product.Description,
+                            BrandId = d.ComboProduct.Product.BrandId,
+                            BrandName = d.ComboProduct.Product.Brand.Name,
+                            CategoryId = d.ComboProduct.Product.CategoryId,
+                            CategoryName = d.ComboProduct.Product.Category.Name,
+                            CreateBy = d.ComboProduct.Product.CreateBy,
+                            CreateDate = d.ComboProduct.Product.CreateDate.HasValue ? d.ComboProduct.Product.CreateDate.Value.ToString("dd-MM-yyyy HH:mm:ss") : null,
+                            UpdateBy = d.ComboProduct.Product.UpdateBy,
+                            UpdateDate = d.ComboProduct.Product.UpdateDate.HasValue ? d.ComboProduct.Product.UpdateDate.Value.ToString("dd-MM-yyyy HH:mm:ss") : null,
+                            DeleteBy = d.ComboProduct.Product.DeleteBy,
+                            DeleteDate = d.ComboProduct.Product.DeleteDate.HasValue ? d.ComboProduct.Product.DeleteDate.Value.ToString("dd-MM-yyyy HH:mm:ss") : null,
+                            DeleteFlag = d.ComboProduct.Product.DeleteFlag,
+                        }).ToList()
                     }).ToList()
                 };
                 if (result == null)
@@ -114,20 +168,42 @@
                     }
                     else
                     {
-                        LogHelper.LogWarning(_logger, "POST", "/api/product-comment", null, null);
+                        LogHelper.LogWarning(_logger, "POST", "/api/cart", null, null);
                         return new ResponseObject(404, "User not found.", null);
                     }
-                    var proCheck = await _unitOfWork.ProductRepository.GetByIdAsync(model.ProductId);
-                    if (proCheck != null)
+                    if (model.ProductId != null)
                     {
-                        cart.ProductId = model.ProductId;
-                        cart.Quantity = model.Quantity;
-                        cart.TotalAmount = model.Quantity * (proCheck.Discount > 0 ? proCheck.Discount : proCheck.Price);
+                        var proCheck = await _unitOfWork.ProductRepository.GetByIdAsync((int)model.ProductId);
+                        if (proCheck != null)
+                        {
+                            cart.Quantity = model.Quantity;
+                            cart.ProductId = model.ProductId;
+                            cart.TotalAmount = model.Quantity * (proCheck.Discount > 0 ? proCheck.Discount : proCheck.Price);
+                        }
+                        else
+                        {
+                            LogHelper.LogWarning(_logger, "POST", "/api/cart", null, null);
+                            return new ResponseObject(404, "Product not found.", null);
+                        }
                     }
-                    else
+                    if (model.ComboId != null)
                     {
-                        LogHelper.LogWarning(_logger, "POST", "/api/product-comment", null, null);
-                        return new ResponseObject(404, "Product not found.", null);
+                        var dataQuery = await _unitOfWork.ComboProductRepository.GetAllAsync(
+                             expression: x => x.ComboId == model.ComboId,
+                            include: query => query.Include(x => x.Combo).Include(x => x.Product)
+                            );
+                        var comboCheck = dataQuery.FirstOrDefault();
+                        if (comboCheck != null)
+                        {
+                            cart.ComboProductId = model.ComboId;
+                            cart.Quantity = model.Quantity;
+                            cart.TotalAmount = model.Quantity * comboCheck.Combo.Price;
+                        }
+                        else
+                        {
+                            LogHelper.LogWarning(_logger, "POST", "/api/cart", null, null);
+                            return new ResponseObject(404, "Product not found.", null);
+                        }
                     }
                     await _unitOfWork.CartRepository.SaveOrUpdateAsync(cart);
                 }
@@ -135,10 +211,24 @@
                 {
                     var cart = cartCheck.FirstOrDefault();
                     cart.Quantity++;
-                    var proCheck = await _unitOfWork.ProductRepository.GetByIdAsync(model.ProductId);
                     cart.ProductId = model.ProductId;
-                    var amountTemp = model.Quantity * (proCheck.Discount > 0 ? proCheck.Discount : proCheck.Price);
-                    cart.TotalAmount += amountTemp;
+                    if(model.ProductId != null)
+                    {
+                        var proCheck = await _unitOfWork.ProductRepository.GetByIdAsync((int)model.ProductId);
+                        var amountTemp = model.Quantity * (proCheck.Discount > 0 ? proCheck.Discount : proCheck.Price);
+                        cart.TotalAmount += amountTemp;
+                    }
+
+                    if (model.ComboId != null)
+                    {
+                        var dataQuery = await _unitOfWork.ComboProductRepository.GetAllAsync(
+                            expression: x => x.ComboId == model.ComboId,
+                           include: query => query.Include(x => x.Combo)
+                           );
+                        var comboCheck = dataQuery.FirstOrDefault();
+                        var amountTemp = model.Quantity * (comboCheck.Combo.Price);
+                        cart.TotalAmount += amountTemp;
+                    }
                     await _unitOfWork.CartRepository.SaveOrUpdateAsync(cart);
                 }
 
@@ -168,27 +258,60 @@
                     return new ResponseObject(400, $"Cart not found with id {id}", null);
                 }
 
-                var product = await _unitOfWork.ProductRepository.GetByIdAsync(cart.ProductId);
-                if (type.Contains(Constant.CART_UPDATE_MINUS))
+                if(cart.ProductId != null && cart.ProductId > 0)
                 {
-                    cart.Quantity--;
-                    if (cart.Quantity == 0)
+                    var product = await _unitOfWork.ProductRepository.GetByIdAsync((int)cart.ProductId);
+                    if (type.Contains(Constant.CART_UPDATE_MINUS))
                     {
-                        LogHelper.LogInformation(_logger, "DELETE", $"/api/cart", null, $"Cart deleted with id {id} as quantity reached 0");
-                        await _unitOfWork.CartRepository.DeleteAsync(cart);
+                        cart.Quantity--;
+                        if (cart.Quantity == 0)
+                        {
+                            LogHelper.LogInformation(_logger, "DELETE", $"/api/cart", null, $"Cart deleted with id {id} as quantity reached 0");
+                            await _unitOfWork.CartRepository.DeleteAsync(cart);
+                        }
+                        else
+                        {
+                            cart.TotalAmount -= (product.Discount > 0 ? product.Discount : product.Price);
+                            await _unitOfWork.CartRepository.SaveOrUpdateAsync(cart);
+                        }
                     }
                     else
                     {
-                        cart.TotalAmount -= (product.Discount > 0 ? product.Discount : product.Price);
+                        cart.Quantity++;
+                        cart.TotalAmount += (product.Discount > 0 ? product.Discount : product.Price);
                         await _unitOfWork.CartRepository.SaveOrUpdateAsync(cart);
                     }
                 }
-                else
+
+                if (cart.ComboProductId != null && cart.ComboProductId > 0)
                 {
-                    cart.Quantity++;
-                    cart.TotalAmount+= (product.Discount > 0 ? product.Discount : product.Price);
-                    await _unitOfWork.CartRepository.SaveOrUpdateAsync(cart);
+                    var dataQuery = await _unitOfWork.ComboProductRepository.GetAllAsync(
+                        expression: x => x.ComboId == cart.ComboProductId,
+                        include: query => query.Include(x => x.Combo)
+                    );
+                    var combo = dataQuery.FirstOrDefault();
+                    if (type.Contains(Constant.CART_UPDATE_MINUS))
+                    {
+                        cart.Quantity--;
+                        if (cart.Quantity == 0)
+                        {
+                            LogHelper.LogInformation(_logger, "DELETE", $"/api/cart", null, $"Cart deleted with id {id} as quantity reached 0");
+                            await _unitOfWork.CartRepository.DeleteAsync(cart);
+                        }
+                        else
+                        {
+                            cart.TotalAmount -= combo.Combo.Price;
+                            await _unitOfWork.CartRepository.SaveOrUpdateAsync(cart);
+                        }
+                    }
+                    else
+                    {
+                        cart.Quantity++;
+                        cart.TotalAmount += combo.Combo.Price;
+                        await _unitOfWork.CartRepository.SaveOrUpdateAsync(cart);
+                    }
                 }
+
                 await _unitOfWork.SaveChangeAsync();
                 LogHelper.LogInformation(_logger, "PUT", "/api/cart", null, cart);
                 return new ResponseObject(200, "Update data successfully", null);
